Refactored database connection and queries to Postges fromat
